Fix: string size is written to 16 bit int (instead of 32 bit)
http://erlang.org/doc/apps/erts/erl_ext_dist.html#id100802

+++ include/bert/types.h
@@ -5,7 +5,7 @@
 
 typedef uint8_t bert_magic_t;
 
-typedef uint32_t bert_string_size_t;
+typedef uint16_t bert_string_size_t;
 typedef uint16_t bert_atom_size_t;
 
 typedef uint32_t bert_bin_size_t;
